Added new DQ_SerialWholeBody class to comply with bug dqrobotics/cpp#18.
The old DQ_WholeBody class is still in the repository for compatibility purposes, but the idea is that it'll become an abstract class in the future and all whole-body classes, including the DQ_SerialWholeBody, will be subclasses of it.

Now the fkm() and pose_jacobian() methods follow the same specification of those same methods in other classes related to serial manipulators. For instance, fkm(q,i) will calculate the fkm up to the ith link in the kinematic chain. Some functions, such as fkm_by_chain(), pose_jacobian_by_chain(), etc., enable to find intermediate values using different indexing. For instance, fkm_by_chain(q,ith) returns the fkm up to the ith subchain; fkm_by_chain(q,ith,jth) returns the fkm up to the jth link of the ith subchain.

Current limitations:

- The documentation must be improved;
- Indexed fkm and related functions do not work for DQ_MobileRobots objects.

Caveat:
- If the first subchain is a mobile robot, the first index of the next subchain is 4, not 2. It happens because the configuration space of a mobile has three dimensions.