This PR adds the property dim_configuration_space in the superclass DQ_Kinematics (#94)
* This commit adds the property dim_configuration_space in DQ_Kinematics and removed it from the subclasses.
* This commit removes the property n_links affecting DQ_SerialManipulator and all its subclasses

[DQ_Kinematics] 
Added the protected property dim_configuration_space.
Updated the credit attributions and the year of the copyright.

[DQ_MobileBase] 
Removed the property dim_configuration_space.
Updated the year of the copyright and Bruno's email.

[DQ_SerialWholeBody] 
Removed the property dim_configuration_space.
Updated the copyright year and Bruno's email.

[DQ_WholeBody] 
Removed the property dim_configuration_space.
Updated the year of the copyright and Bruno's email.

[DQ_FreeFlyingRobot] 
Removed the property dim_configuration_space.
Updated the year of the copyright.

[DQ_SerialManipulator] 
Removed n_links. Now the class uses dim_configuration_space from DQ_Kinematics. 
Updated the credit attributions and the year of the copyright.

[DQ_SerialManipulatorDH, MDH] Removed the definition of n_links to comply with the changes in DQ_SerialManipulator and DQ_Kinematics.